+++ Options/Locales/ruRU.lua
@@ -14,6 +14,7 @@ local L = _Locale("ruRU")
 
 if not L then return end
 
+-- L.DESCRIPTION = "Description"
 -- L.HIDE_COMPLETED_OBJECTIVES = "Hide completed objectives"
 
 L.ACHIEVEMENT = "Достижение"